+++ lib/oasis/entities/blob.rb
@@ -1,0 +1,2 @@
+class Blob < Hanami::Entity
+end