+++ lib/oasis/repositories/blob_repository.rb
@@ -1,0 +1,2 @@
+class BlobRepository < Hanami::Repository
+end