+++ spec/oasis/repositories/blob_repository_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+describe BlobRepository do
+  # place your tests here
+end