+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Add Blob entities to represent data and metadata
+* Add a search endpoint on data
 * Add project to Travis CI
 * Add `README.md`, `LICENSE.md`, and `CHANGELOG.md` files.
 * Boot strap the specs (Guard, Minitest reports, helpers)